Django/rest-blog/pyproject.toml
[project]
name = "rest-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "uuuj0821",email = "uuuj0821@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "ipython (>=9.0.2,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
